+++ lib/core/zenith/zenith_core.dart
@@ -1,4 +1,5 @@
 export 'async_value.dart';
+export 'zenith_concurrency.dart';
 export 'zenith_container.dart';
 export 'zenith_key.dart';
 export 'zenith_node.dart';